Refactor OrdersTable component files structure
