+++ src/cow-react/modules/limitOrders/pure/Orders/OrdersTable.tsx
@@ -1,10 +1,85 @@
 import { Order, OrderStatus } from 'state/orders/actions'
-import { CurrencyAmount, Fraction } from '@uniswap/sdk-core'
+import { Currency, CurrencyAmount, Fraction } from '@uniswap/sdk-core'
 import { Trans } from '@lingui/macro'
-import * as styledEl from './OrdersTable.styled'
 import { useEffect, useState } from 'react'
 import { OrdersTablePagination } from './OrdersTablePagination'
 import { formatSmart } from 'utils/format'
+import styled from 'styled-components/macro'
+import CurrencyLogo from 'components/CurrencyLogo'
+
+const statusColorMap: { [key in OrderStatus]: string } = {
+  [OrderStatus.PENDING]: '#badbe8',
+  [OrderStatus.PRESIGNATURE_PENDING]: '#badbe8',
+  [OrderStatus.EXPIRED]: '#eeaaaa',
+  [OrderStatus.FULFILLED]: '#d5eab3',
+  [OrderStatus.CANCELLED]: '#fcecb4',
+  // TODO: decide what color for each state
+  [OrderStatus.CREATING]: '#badbe8',
+  [OrderStatus.REFUNDED]: '#eeaaaa',
+  [OrderStatus.REFUNDING]: '#eeaaaa',
+  [OrderStatus.REJECTED]: '#eeaaaa',
+}
+
+const TableBox = styled.div`
+  display: block;
+  border-radius: 16px;
+  border: 2px solid ${({ theme }) => theme.border2};
+`
+
+const Header = styled.div`
+  display: grid;
+  grid-gap: 10px;
+  grid-template-columns: repeat(4, minmax(0, 1fr));
+  align-items: center;
+  border-bottom: 2px solid ${({ theme }) => theme.border2};
+
+  > div {
+    padding: 12px;
+    overflow: hidden;
+    font-size: 14px;
+  }
+`
+
+const Rows = styled.div`
+  display: block;
+`
+
+const Row = styled(Header)`
+  :last-child {
+    border-bottom: 0;
+  }
+`
+
+const StatusItem = styled.div<{ status: OrderStatus; cancelling: boolean }>`
+  display: inline-block;
+  background: ${({ status, cancelling }) => (cancelling ? statusColorMap.cancelled : statusColorMap[status])};
+  color: ${({ theme }) => theme.text2};
+  padding: 5px 10px;
+  border-radius: 3px;
+`
+
+const AmountItem = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  white-space: nowrap;
+`
+
+const RateValue = styled.span`
+  font-size: 12px;
+`
+
+export function CurrencyAmountItem({ amount }: { amount: CurrencyAmount<Currency> }) {
+  return (
+    <AmountItem title={amount.toExact() + ' ' + amount.currency.symbol}>
+      <div>
+        <CurrencyLogo currency={amount.currency} size="24px" />
+      </div>
+      <span>{formatSmart(amount)}</span>
+      <span>{amount.currency.symbol}</span>
+    </AmountItem>
+  )
+}
 
 export interface OrdersTableProps {
   orders: Order[]
@@ -36,8 +111,8 @@ export function OrdersTable({ orders }: OrdersTableProps) {
 
   return (
     <>
-      <styledEl.TableBox>
-        <styledEl.Header>
+      <TableBox>
+        <Header>
           <div>
             <Trans>Sell</Trans>
           </div>
@@ -50,39 +125,39 @@ export function OrdersTable({ orders }: OrdersTableProps) {
           <div>
             <Trans>Status</Trans>
           </div>
-        </styledEl.Header>
-        <styledEl.Rows>
+        </Header>
+        <Rows>
           {ordersPage.map((order) => {
             const sellAmount = CurrencyAmount.fromRawAmount(order.inputToken, order.sellAmount.toString())
             const buyAmount = CurrencyAmount.fromRawAmount(order.outputToken, order.buyAmount.toString())
             const price = new Fraction(order.buyAmount.toString(), order.sellAmount.toString())
 
             return (
-              <styledEl.Row key={order.id}>
+              <Row key={order.id}>
                 <div>
-                  <styledEl.CurrencyAmountItem amount={sellAmount} />
+                  <CurrencyAmountItem amount={sellAmount} />
                 </div>
                 <div>
-                  <styledEl.CurrencyAmountItem amount={buyAmount} />
+                  <CurrencyAmountItem amount={buyAmount} />
                 </div>
                 <div>
-                  <styledEl.RateValue>
+                  <RateValue>
                     1 {order.inputToken.symbol} ={' '}
                     <span title={price.toSignificant(18) + ' ' + order.outputToken.symbol}>
                       {formatSmart(price)} {order.outputToken.symbol}
                     </span>
-                  </styledEl.RateValue>
+                  </RateValue>
                 </div>
                 <div>
-                  <styledEl.StatusItem cancelling={!!order.isCancelling} status={order.status}>
+                  <StatusItem cancelling={!!order.isCancelling} status={order.status}>
                     {order.isCancelling ? 'Cancelling...' : orderStatusTitleMap[order.status]}
-                  </styledEl.StatusItem>
+                  </StatusItem>
                 </div>
-              </styledEl.Row>
+              </Row>
             )
           })}
-        </styledEl.Rows>
-      </styledEl.TableBox>
+        </Rows>
+      </TableBox>
       <OrdersTablePagination
         pageSize={pageSize}
         totalCount={orders.length}
