--- src/cow-react/modules/limitOrders/pure/Orders/styled.ts
@@ -1,17 +1,0 @@
-import styled from 'styled-components/macro'
-import { Widget } from '../Widget'
-
-export const Orders = styled(Widget)`
-  min-height: 200px;
-  width: 100%;
-`
-
-export const OrdersTitle = styled.h3`
-  margin: 0 0 20px 0;
-`
-
-export const EmptyOrdersMessage = styled.p`
-  margin: 10px 20px;
-  font-size: 18px;
-  font-weight: 600;
-`
